chore: resolve ts usage
