+++ sdk/dir-js/.prettierrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('gts/.prettierrc.json')
+}
