--- sdk/dir-js/client/index.ts
@@ -1,4 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './client';
