--- sdk/dir-js/models/index.ts
@@ -1,8 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-export * as core_v1 from './core_v1';
-export * as routing_v1 from './routing_v1';
-export * as search_v1 from './search_v1';
-export * as sign_v1 from './sign_v1';
-export * as store_v1 from './store_v1';
