--- sdk/dir-js/models/routing_v1.ts
@@ -1,9 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/routing/v1/peer_pb';
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/routing/v1/publication_service_pb';
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/routing/v1/publication_service_pb.client';
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/routing/v1/record_query_pb';
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/routing/v1/routing_service_pb';
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/routing/v1/routing_service_pb.client';
