--- sdk/dir-js/models/sign_v1.ts
@@ -1,5 +1,0 @@
-// Copyright AGNTCY Contributors (https://github.com/agntcy)
-// SPDX-License-Identifier: Apache-2.0
-
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/sign/v1/sign_service_pb';
-export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/sign/v1/sign_service_pb.client';
