+++ sdk/dir-js/src/index.ts
@@ -1,5 +1,5 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './client';
 export * as models from './models';
+export * from './client/client';
