+++ sdk/dir-js/src/models/search_v1.ts
@@ -1,5 +1,6 @@
 // Copyright AGNTCY Contributors (https://github.com/agntcy)
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '@buf/agntcy_dir.bufbuild_es/agntcy/dir/search/v1/record_query_pb';
-export * from '@buf/agntcy_dir.bufbuild_es/agntcy/dir/search/v1/search_service_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/search/v1/record_query_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/search/v1/search_service_pb';
+export * from '@buf/agntcy_dir.community_timostamm-protobuf-ts/agntcy/dir/search/v1/search_service_pb.client';
