+++ sdk/dir-js/test/client.test.ts
@@ -2,14 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {describe, test, beforeAll, afterAll, expect} from 'vitest';
-
 import {execSync} from 'node:child_process';
 import {readFileSync, rmSync} from 'node:fs';
-import {env} from 'node:process';
-import {create} from '@bufbuild/protobuf';
+import process from 'node:process';
 
 import {validate as isValidUUID} from 'uuid';
 import {v4 as uuidv4} from 'uuid';
+import {Struct} from '@buf/agntcy_dir.community_timostamm-protobuf-ts/google/protobuf/struct_pb';
 
 import {Client, Config, models} from '../src';
 
@@ -26,41 +25,39 @@ function genRecords(
 ): models.core_v1.Record[] {
   const records: models.core_v1.Record[] = [];
   for (let index = 0; index < count; index++) {
-    records.push(
-      create(models.core_v1.RecordSchema, {
-        data: {
-          name: `agntcy-${testFunctionName}-${index}-${uuidv4().substring(0, 8)}`,
-          version: 'v3.0.0',
-          schema_version: 'v0.7.0',
-          description: "Research agent for Cisco's marketing strategy.",
-          authors: ['Cisco Systems'],
-          created_at: '2025-03-19T17:06:37Z',
-          skills: [
-            {
-              name: 'natural_language_processing/natural_language_generation/text_completion',
-              id: 10201,
-            },
-            {
-              name: 'natural_language_processing/analytical_reasoning/problem_solving',
-              id: 10702,
-            },
-          ],
-          locators: [
-            {
-              type: 'docker-image',
-              url: 'https://ghcr.io/agntcy/marketing-strategy',
-            },
-          ],
-          domains: [
-            {
-              name: 'technology/networking',
-              id: 103,
-            },
-          ],
-          modules: [],
-        },
+    records.push({
+      data: Struct.fromJson({
+        name: `agntcy-${testFunctionName}-${index}-${uuidv4().substring(0, 8)}`,
+        version: 'v3.0.0',
+        schema_version: 'v0.7.0',
+        description: "Research agent for Cisco's marketing strategy.",
+        authors: ['Cisco Systems'],
+        created_at: '2025-03-19T17:06:37Z',
+        skills: [
+          {
+            name: 'natural_language_processing/natural_language_generation/text_completion',
+            id: 10201,
+          },
+          {
+            name: 'natural_language_processing/analytical_reasoning/problem_solving',
+            id: 10702,
+          },
+        ],
+        locators: [
+          {
+            type: 'docker-image',
+            url: 'https://ghcr.io/agntcy/marketing-strategy',
+          },
+        ],
+        domains: [
+          {
+            name: 'technology/networking',
+            id: 103,
+          },
+        ],
+        modules: [],
       }),
-    );
+    });
   }
 
   return records;
@@ -71,7 +68,7 @@ describe('Client', () => {
 
   beforeAll(async () => {
     // Verify that DIRCTL_PATH is set in the environment
-    expect(env.DIRCTL_PATH).toBeDefined();
+    expect(process.env.DIRCTL_PATH).toBeDefined();
 
     // Initialize the client
     client = new Client(Config.loadFromEnv());
@@ -91,7 +88,7 @@ describe('Client', () => {
     expect(recordRefs).toHaveLength(2);
 
     for (const ref of recordRefs) {
-      expect(ref).toBeTypeOf(typeof models.core_v1.RecordRefSchema);
+      expect(ref).toBeInstanceOf(models.core_v1.RecordRef);
       expect(ref.cid).toHaveLength(59);
     }
   });
@@ -107,7 +104,7 @@ describe('Client', () => {
 
     for (let index = 0; index < pulledRecords.length; index++) {
       const record = pulledRecords[index];
-      expect(record).toBeTypeOf(typeof models.core_v1.RecordSchema);
+      expect(record).toBeInstanceOf(models.core_v1.Record);
       expect(record).toEqual(records[index]);
     }
   });
@@ -116,15 +113,15 @@ describe('Client', () => {
     const records = genRecords(1, 'search');
     await client.push(records);
 
-    const searchRequest = create(models.search_v1.SearchRequestSchema, {
-      queries: [
-        {
-          type: models.search_v1.RecordQueryType.SKILL_ID,
-          value: '10201',
-        },
-      ],
+    const searchQuery: models.search_v1.RecordQuery = {
+      type: models.search_v1.RecordQueryType.SKILL_ID,
+      value: '10201',
+    };
+
+    const searchRequest: models.search_v1.SearchRequest = {
+      queries: [searchQuery],
       limit: 2,
-    });
+    };
 
     const objects = await client.search(searchRequest);
 
@@ -133,7 +130,7 @@ describe('Client', () => {
     expect(objects.length).toBeGreaterThan(0);
 
     for (const obj of objects) {
-      expect(obj).toBeTypeOf(typeof models.search_v1.SearchResponseSchema);
+      expect(obj).toBeInstanceOf(models.search_v1.SearchResponse);
     }
   });
 
@@ -147,7 +144,7 @@ describe('Client', () => {
     expect(metadatas).toHaveLength(2);
 
     for (const metadata of metadatas) {
-      expect(metadata).toBeTypeOf(typeof models.core_v1.RecordMetaSchema);
+      expect(metadata).toBeInstanceOf(models.core_v1.RecordMeta);
     }
   });
 
@@ -155,16 +152,14 @@ describe('Client', () => {
     const records = genRecords(1, 'publish');
     const recordRefs = await client.push(records);
 
-    await client.publish(
-      create(models.routing_v1.PublishRequestSchema, {
-        request: {
-          case: 'recordRefs',
-          value: {
-            refs: recordRefs,
-          },
+    await client.publish({
+      request: {
+        oneofKind: 'recordRefs',
+        recordRefs: {
+          refs: recordRefs,
         },
-      }),
-    );
+      },
+    });
   });
 
   test('list', async () => {
@@ -172,68 +167,59 @@ describe('Client', () => {
     const recordRefs = await client.push(records);
 
     // Publish records
-    await client.publish(
-      create(models.routing_v1.PublishRequestSchema, {
-        request: {
-          case: 'recordRefs',
-          value: {
-            refs: recordRefs,
-          },
+    await client.publish({
+      request: {
+        oneofKind: 'recordRefs',
+        recordRefs: {
+          refs: recordRefs,
         },
-      }),
-    );
+      },
+    });
 
     // Sleep to allow the publication to be indexed
     await new Promise(resolve => setTimeout(resolve, 5000));
 
     // Query for records in the domain
-    const objects = await client.list(
-      create(models.routing_v1.ListRequestSchema, {
-        queries: [
-          {
-            type: models.routing_v1.RecordQueryType.DOMAIN,
-            value: 'technology/networking',
-          },
-        ],
-      }),
-    );
+    const query: models.routing_v1.RecordQuery = {
+      type: models.routing_v1.RecordQueryType.DOMAIN,
+      value: 'technology/networking',
+    };
+    const objects = await client.list({
+      queries: [query],
+    });
 
     expect(objects).not.toBeNull();
     expect(objects).toBeInstanceOf(Array);
     expect(objects.length).not.toBe(0);
 
     for (const obj of objects) {
-      expect(obj).toBeTypeOf(typeof models.routing_v1.ListResponseSchema);
+      expect(obj).toBeInstanceOf(models.routing_v1.ListResponse);
     }
-  }, 30000);
+  });
 
   test('unpublish', async () => {
     const records = genRecords(1, 'unpublish');
     const recordRefs = await client.push(records);
 
     // Publish records
-    await client.publish(
-      create(models.routing_v1.PublishRequestSchema, {
-        request: {
-          case: 'recordRefs',
-          value: {
-            refs: recordRefs,
-          },
+    await client.publish({
+      request: {
+        oneofKind: 'recordRefs',
+        recordRefs: {
+          refs: recordRefs,
         },
-      }),
-    );
+      },
+    });
 
     // Unpublish
-    await client.unpublish(
-      create(models.routing_v1.UnpublishRequestSchema, {
-        request: {
-          case: 'recordRefs',
-          value: {
-            refs: recordRefs,
-          },
+    await client.unpublish({
+      request: {
+        oneofKind: 'recordRefs',
+        recordRefs: {
+          refs: recordRefs,
         },
-      }),
-    );
+      },
+    });
   });
 
   test('delete', async () => {
@@ -251,13 +237,13 @@ describe('Client', () => {
       (
         recordRef: models.core_v1.RecordRef,
       ): models.store_v1.PushReferrerRequest => {
-        return create(models.store_v1.PushReferrerRequestSchema, {
+        return {
           recordRef: recordRef,
           options: {
-            case: 'signature',
-            value: {},
+            oneofKind: 'signature',
+            signature: {} as models.sign_v1.Signature,
           },
-        });
+        };
       },
     );
 
@@ -266,7 +252,7 @@ describe('Client', () => {
     expect(response).toHaveLength(2);
 
     for (const r of response) {
-      expect(r).toBeTypeOf(typeof models.store_v1.PushReferrerResponseSchema);
+      expect(r).toBeInstanceOf(models.store_v1.PushReferrerResponse);
     }
   });
 
@@ -278,13 +264,13 @@ describe('Client', () => {
       (
         recordRef: models.core_v1.RecordRef,
       ): models.store_v1.PullReferrerRequest => {
-        return create(models.store_v1.PullReferrerRequestSchema, {
+        return {
           recordRef: recordRef,
           options: {
-            case: 'pullSignature',
-            value: true,
+            oneofKind: 'pullSignature',
+            pullSignature: true,
           },
-        });
+        };
       },
     );
 
@@ -293,7 +279,7 @@ describe('Client', () => {
     expect(response).toHaveLength(2);
 
     for (const r of response) {
-      expect(r).toBeTypeOf(typeof models.store_v1.PullReferrerResponseSchema);
+      expect(r).toBeInstanceOf(models.store_v1.PullReferrerResponse);
     }
   });
 
@@ -301,7 +287,7 @@ describe('Client', () => {
     const records = genRecords(2, 'sign_verify');
     const recordRefs = await client.push(records);
 
-    const shellEnv = {...env};
+    const shellEnv = {...process.env};
     const keyPassword = 'testing-key';
 
     // Clean up any existing keys
@@ -310,7 +296,7 @@ describe('Client', () => {
 
     try {
       // Generate key pair
-      const cosignPath = env['COSIGN_PATH'] || 'cosign';
+      const cosignPath = process.env['COSIGN_PATH'] || 'cosign';
       execSync(`${cosignPath} generate-key-pair`, {
         env: {...shellEnv, COSIGN_PASSWORD: keyPassword},
         encoding: 'utf8',
@@ -324,25 +310,25 @@ describe('Client', () => {
       const clientId = shellEnv['OIDC_CLIENT_ID'] || 'sigstore';
 
       // Create signing providers
-      const keyRequest = create(models.sign_v1.SignRequestSchema, {
+      const keyRequest: models.sign_v1.SignRequest = {
         recordRef: recordRefs[0],
         provider: {
           request: {
-            case: 'key',
-            value: {
+            oneofKind: 'key',
+            key: {
               privateKey: keyFile,
               password: Buffer.from(keyPassword, 'utf-8'),
             },
           },
         },
-      });
+      };
 
-      const oidcRequest = create(models.sign_v1.SignRequestSchema, {
+      const oidcRequest: models.sign_v1.SignRequest = {
         recordRef: recordRefs[1],
         provider: {
           request: {
-            case: 'oidc',
-            value: {
+            oneofKind: 'oidc',
+            oidc: {
               idToken: token,
               options: {
                 oidcProviderUrl: providerUrl,
@@ -350,42 +336,43 @@ describe('Client', () => {
             },
           },
         },
-      });
+      };
 
       // Sign test
-      client.sign(keyRequest);
-      client.sign(oidcRequest, clientId);
+      const keyCommandResult = client.sign(keyRequest);
+      expect(keyCommandResult.signature).toBeDefined();
+
+      const oidcCommandResult = client.sign(oidcRequest, clientId);
+      expect(oidcCommandResult.signature).toBeDefined();
 
       // Verify test
       for (const ref of recordRefs) {
-        const response = await client.verify(
-          create(models.sign_v1.VerifyRequestSchema, {
-            recordRef: ref,
-          }),
-        );
+        const response = await client.verify({
+          recordRef: ref,
+        });
         expect(response.success).toBe(true);
       }
 
       // Test invalid CID
-      try {
-        client.sign(
-          create(models.sign_v1.SignRequestSchema, {
-            recordRef: {cid: 'invalid-cid'},
-            provider: {
-              request: {
-                case: 'key',
-                value: {
-                  privateKey: Uint8Array.from([]),
-                  password: Uint8Array.from([]),
-                },
-              },
+      const invalidRequest: models.sign_v1.SignRequest = {
+        recordRef: {cid: 'invalid-cid'},
+        provider: {
+          request: {
+            oneofKind: 'key',
+            key: {
+              privateKey: Uint8Array.from([]),
+              password: Uint8Array.from([]),
             },
-          }),
-        );
+          },
+        },
+      };
+
+      try {
+        client.sign(invalidRequest);
         expect.fail('Should have thrown error for invalid CID');
       } catch (error) {
         if (error instanceof Error) {
-          expect(error.message).toContain('failed to decode CID invalid-cid');
+          expect(error.message).toContain('Failed to sign the object');
         }
       }
     } catch (error) {
@@ -399,46 +386,35 @@ describe('Client', () => {
 
   test('sync', async () => {
     // Create sync
-    const createResponse = await client.create_sync(
-      create(models.store_v1.CreateSyncRequestSchema, {
-        remoteDirectoryUrl:
-          env['DIRECTORY_SERVER_PEER1_ADDRESS'] || '0.0.0.0:8891',
-      }),
-    );
-    expect(createResponse).toBeTypeOf(
-      typeof models.store_v1.CreateSyncResponseSchema,
-    );
+    const createResponse = await client.create_sync({
+      remoteDirectoryUrl:
+        process.env['DIRECTORY_SERVER_PEER1_ADDRESS'] || '0.0.0.0:8891',
+    });
+    expect(createResponse).toBeInstanceOf(models.store_v1.CreateSyncResponse);
 
     const syncId = createResponse.syncId;
     expect(isValidUUID(syncId)).toBe(true);
 
     // List syncs
-    const listResponse = await client.list_syncs(
-      create(models.store_v1.ListSyncsRequestSchema, {}),
-    );
+    const listRequest: models.store_v1.ListSyncsRequest = {};
+    const listResponse = await client.list_syncs(listRequest);
     expect(listResponse).toBeInstanceOf(Array);
 
     for (const syncItem of listResponse) {
-      expect(syncItem).toBeTypeOf(typeof models.store_v1.ListSyncsItemSchema);
+      expect(syncItem).toBeInstanceOf(models.store_v1.ListSyncsItem);
       expect(isValidUUID(syncItem.syncId)).toBe(true);
     }
 
     // Get sync
-    const getResponse = await client.get_sync(
-      create(models.store_v1.GetSyncRequestSchema, {
-        syncId: syncId,
-      }),
-    );
-    expect(getResponse).toBeTypeOf(
-      typeof models.store_v1.GetSyncResponseSchema,
-    );
+    const getResponse = await client.get_sync({
+      syncId: syncId,
+    });
+    expect(getResponse).toBeInstanceOf(models.store_v1.GetSyncResponse);
     expect(getResponse.syncId).toEqual(syncId);
 
     // Delete sync
-    await client.delete_sync(
-      create(models.store_v1.DeleteSyncRequestSchema, {
-        syncId: syncId,
-      }),
-    );
+    await client.delete_sync({
+      syncId: syncId,
+    });
   });
 });
