+++ sdk/dir-js/tsconfig.json
@@ -1,30 +1,19 @@
 {
+  "extends": "./node_modules/gts/tsconfig-google.json",
   "compilerOptions": {
-    "target": "es2018",
-    "module": "esnext",
-    "lib": ["es2018"],
+    "rootDir": ".",
     "outDir": "./dist",
-    "types": ["node"],
-    "rootDir": "./",
-    "strict": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
     "declaration": true,
-    "declarationMap": true,
-    "sourceMap": false,
-    "resolveJsonModule": true,
+    "strict": true,
     "moduleResolution": "node",
-    "noEmitOnError": true,
-    "isolatedModules": true
   },
   "include": [
     "**/*.ts",
-    "**/tests/**/*.test.ts"
+    "test/**/*.ts"
   ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
+  "ts-node": {
+    "compilerOptions": {
+      "module": "ES2020"
+    }
+  }
 }
